complete UDP ,more PC SIM msg,latency test script
-Complete 11-15 UDP Message
-Complete FC Simulator Msg
-Add Latency Test Script

+++ arduino/mavlink_gateway/custom_protocol.h
@@ -23,7 +23,7 @@ static uint8_t  gs_arm;
 static uint8_t gs_fix_type;
 static uint16_t gs_hdop;
 static uint8_t gs_num_sat;
-static uint8_t gs_volt_batt;
+static uint16_t gs_volt_batt;
 
 //Msg ID 13
 static float gs_roll;
